+++ R/constants.R
@@ -1,0 +1,29 @@
+#' Output constants
+#'
+#' @name constants
+#' @rdname constants
+#'
+#' @keywords constants
+RESULT_NAMES <- c(
+  "parameters",
+  "costs",
+  "time_series",
+  "interventions",
+  "capacities",
+  "gdp",
+  "vsl"
+)
+
+#' @name constants
+#'
+#' @keywords constants
+TIMESERIES_NAMES <- c(
+  "prevalence",
+  "hospitalised",
+  "dead",
+  "vaccinated",
+  "new_infected",
+  "new_hospitalised",
+  "new_dead",
+  "new_vaccinated"
+)