+++ tests/testthat/test_model_run.R
@@ -38,13 +38,20 @@ test_that("can run model and return results", {
     mock_new_vaccinations_result
   )
 
+  # Mock output of `model_run()`
   mock_costs_data <- daedalus_mock_costs()
+
   mock_get_costs <- mockery::mock(mock_costs_data)
   mockery::stub(model_run, "daedalus::get_costs", mock_get_costs)
+
   mock_get_gdp <- mockery::mock(9999)
   mockery::stub(model_run, "get_annual_gdp", mock_get_gdp)
+
   mock_get_average_vsl <- mockery::mock(10000)
   mockery::stub(model_run, "get_average_vsl", mock_get_average_vsl)
+
+  mock_get_age_vsl <- mockery::mock(c(500, 1000, 1500, 2000))
+  mockery::stub(model_run, "get_age_vsl", mock_get_age_vsl)
 
   ctx <- "CAN"
   hosp_cap <- 4500
@@ -78,29 +85,13 @@ test_that("can run model and return results", {
 
   expect_named(
     res,
-    c(
-      "parameters",
-      "costs",
-      "time_series",
-      "interventions",
-      "capacities",
-      "gdp",
-      "average_vsl"
-    )
+    RESULT_NAMES
   )
   expect_named(
     res$time_series,
-    c(
-      "prevalence",
-      "hospitalised",
-      "dead",
-      "vaccinated",
-      "new_infected",
-      "new_hospitalised",
-      "new_dead",
-      "new_vaccinated"
-    )
+    TIMESERIES_NAMES
   )
+
   expect_identical(res$time_series$prevalence, c(28L, 87L))
   expect_identical(res$time_series$hospitalised, c(11L, 31L))
   expect_identical(res$time_series$dead, c(15L, 37L))
@@ -113,6 +104,7 @@ test_that("can run model and return results", {
     mock_new_vaccinations_result$new_vaccinations
   )
   expect_identical(res$parameters, parameters)
+
   expect_nested_mock_costs(res$costs)
   expect_identical(
     res$interventions,
@@ -128,5 +120,14 @@ test_that("can run model and return results", {
   expect_identical(res$capacities[[1]]$id, "hospital_capacity")
   expect_identical(res$capacities[[1]]$value, 4500)
   expect_identical(res$gdp, 9999)
-  expect_identical(res$average_vsl, 10000)
+  expect_identical(
+    res$vsl,
+    list(
+      average = 10000,
+      pre_school = 500,
+      school_age = 1000,
+      working_age = 1500,
+      retirement_age = 2000
+    )
+  )
 })