+++ NEWS.md
@@ -1,3 +1,19 @@
+# daedalus.api 0.1.6
+
+This patch version updates cost and results outputs (PR #47).
+
+- Updates to function `get_nested_costs()` and JSON schema `scenarioCosts.json` to return costs as an array of metric-value pairs; this allows costs to be expressed in different units.
+
+- Updates to results schema `scenarioResults.json` to return the age-specific VSL for countries.
+
+- Updates to metadata schema to return metric and value for costs with multiple units.
+
+- Added results list names and time-series names as package constants.
+
+- Moved `cost_item()` function from inline helper to internal package function.
+
+- Updates to tests helper functions; added helper tests for nested list outputs in `R/tests.R`.
+
 # daedalus.api 0.1.5
 
 - Update `model_run()` daedalus model runner to error if a country hospital capacity <= 0 is passed as this reportedly causes run errors;